Significant changes to user_flair scraper
Changed the scraper so that it saves every 1000 unique user/flair combinations to a .csv so that the .csv is gradually filled up. This reduces the memory use of the script and also safeguards against the possibility that the script is terminated whilst still running. Now, if this happens, at most 1000 observations are lost rather than all of them.