Substantial changes to the user flair scraper
The scraper now looks through the top posts of all time rather than the most recent posts. This is because PRAW limits the posts that can be looped through to 1000. Since the top 1000 posts have more comments than the 1000 most recent posts, it makes sense to look through these (they are also likely to be further apart in time and therefore have less overlap in terms of the user commenting). I also got rid of the loop breaker since it's unnecessary: the script will end once each of the 1000 top posts has been looped through.